// colorbar/types.ts
/**
 * @fileoverview Types for the color bar.
 *
 * @license
 * Copyright 2024 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import {TextStyle} from '../text/text_style';

/**
 * A marker is a point on the color bar.
 */
export interface Marker {
  value: number;
}

/**
 * Orientation in which to draw the color bar.
 * Color bar orientation.
 */
export enum Orientation {
  HORIZONTAL = 'horizontal',
  VERTICAL = 'vertical',
}

/**
 * The width is always the size of the color bar base line (the length of the
 * color gradient).
 */
export interface Options {
  top: number;
  left: number;
  width: number;
  height: number;
  orientation: Orientation;
  textStyle: TextStyle;
  markerColor: string;
  numberFormat: string | null;
}